Add files via upload
#include<iostream>:-This line includes the standard input/output library in the program. It allows you to use input and output functions like cin and cout.
using namespace std means that we can use names for objects and variables from the standard library.
The base class is named Vehicle.It contains a public method vehicle() which prints "I am a vehicle".
twowheeler is a class derived from the Vehicle class.It inherits all public members of the Vehicle class.It contains a public method twoWheeler() which prints "I am a twowheeler".
bike is a class derived from the twowheeler class. It inherits all public members of the twowheeler class, and also from Vehicle. It contains a public method Bike() which prints "I am a bike".
int main() is the main function where the execution of the program begins.
bike mybike; creates an object of the bike class named mybike.
mybike.Bike(); calls the Bike() method of the bike class, which prints "I am a bike".
mybike.twoWheeler(); calls the twoWheeler() method of the twowheeler class, which prints "I am a twowheeler".
mybike.vehicle(); calls the vehicle() method of the Vehicle class, which prints "I am a vehicle".